init container builds only used plugins

+++ build_plugins.sh
@@ -2,6 +2,19 @@
 
 set -e
 
+shallBuild () {
+  SHALL_BUILD="NO"
+
+  [ -f $VERADCO_CONF ] || SHALL_BUILD="YES"
+
+  if [ "$SHALL_BUILD" != "YES" ]; then
+    PLUGPATH=$(echo "$1" | sed "s#/release/#/app/#")
+    IS_IN_CONF=$(cat $VERADCO_CONF | yq ".plugins[] | select(.path==\"$PLUGPATH\") | .name")
+    SHALL_BUILD="YES"
+    [ "$IS_IN_CONF" != "" ] || SHALL_BUILD="NO"
+  fi
+}
+
 DEFAULT_PLUGINS_PATH="/go/src/built-in_plugins"
 
 [ -z "$PLUGINS_PATH" ] && PLUGINS_PATH=$DEFAULT_PLUGINS_PATH
@@ -14,19 +27,28 @@ cd $PLUGINS_PATH
 
 PLUGINS_LIB_PATH="/release/plugins"
 
+[ -z "$VERADCO_CONF" ] && VERADCO_CONF="/conf/veradco.yaml"
+
 mkdir -p "$PLUGINS_LIB_PATH"
 
 for folder in $(ls -d $PLUGINS_PATH/*/); do
   cd "$folder"
   plugin_name=${PLUGIN_PREFIX}$(echo -n "$folder" | grep -o -E "[^/]+/$" | grep -o -E "^[^/]+")
-  echo "Building plugin $plugin_name"
-  set +e
-  rm go.mod go.sum
-  set -e
-  go mod init "github.com/smart-duck/veradco/$plugin_name"
-  go mod edit -replace github.com/smart-duck/veradco=../../veradco
-  go mod tidy
-  go build -buildmode=plugin -o "$PLUGINS_LIB_PATH/$plugin_name.so" plug.go
+
+  shallBuild "$PLUGINS_LIB_PATH/$plugin_name.so"
+
+  if [ "$SHALL_BUILD" = "YES" ]; then
+    echo "Building plugin $plugin_name"
+    set +e
+    rm go.mod go.sum
+    set -e
+    go mod init "github.com/smart-duck/veradco/$plugin_name"
+    go mod edit -replace github.com/smart-duck/veradco=../../veradco
+    go mod tidy
+    go build -buildmode=plugin -o "$PLUGINS_LIB_PATH/$plugin_name.so" plug.go
+  else
+    echo "NO NEED TO BUILD plugin $plugin_name"
+  fi
   cd ..
 done
